# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  my-web-app_com_smart-logistic-hub_internal_warehouse_dto.CreateWarehouseRequest:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      location:
        $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.Location'
      manager_name:
        type: string
      name:
        type: string
      warehouse_code:
        type: string
    required:
    - address
    - name
    - warehouse_code
    type: object
  my-web-app_com_smart-logistic-hub_internal_warehouse_dto.Location:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  my-web-app_com_smart-logistic-hub_internal_warehouse_dto.PaginatedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.WarehouseResponse'
        type: array
      limit:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_warehouse_dto.UpdateWarehouseRequest:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      is_active:
        type: boolean
      location:
        $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.Location'
      manager_name:
        type: string
      name:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_warehouse_dto.WarehouseResponse:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      manager_name:
        type: string
      name:
        type: string
      updated_at:
        type: string
      warehouse_code:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: Smart Logistics Hub backend API. Authentication uses the "Authorization"
    header with format "Bearer &lt;token&gt;" (JWT issued by Keycloak).
  title: Smart Logistics Hub API
  version: 1.0.0
paths:
...
  /warehouses:
    get:
      parameters:
      - default: 0
        description: Number of items to skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Max items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List warehouses
      tags:
      - Warehouses
    post:
      consumes:
      - application/json
      parameters:
      - description: Warehouse payload
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.CreateWarehouseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.WarehouseResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a warehouse
      tags:
      - Warehouses
  /warehouses/{id}:
    delete:
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a warehouse
      tags:
      - Warehouses
    get:
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.WarehouseResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a warehouse by ID
      tags:
      - Warehouses
    patch:
      consumes:
      - application/json
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: integer
      - description: Warehouse update
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.UpdateWarehouseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_warehouse_dto.WarehouseResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a warehouse
      tags:
      - Warehouses
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"